+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@ It features the following interaction:
     * ``version``: the REST API version and release date.
     * ``test``: next word to test.
     * ``tomorrow_test_count``: number of tests for the next day.
+    * ``phonetic_reading``: phonetic reading of a text.
   * On POST, ``action_type`` can be:
     * ``simterms``: similar terms to a given term
     * ``add_translation``: add a translation for a new word
@@ -64,6 +65,7 @@ the intendended page. Same goes for "New Text" on "Archived Texts" page.
   * ``inc/ajax_add_term_transl.php``, use ``action_type=add_translation`` or ``action_type=update_translation``. The arguments were also changed.
   * ``inc/ajax_check_regexp.php`` should be accessed through ``action_type=regexp``. Argument ``regex`` is now ``regexp``, on post only.
   * ``inc/ajax_chg_term_status.php`` should be accessed through ``action_type=increase_term_status``. Argument ``data`` is now ``status_up``, on post only.
+  * ``inc/ajax_get_phonetic.php`` should be accessed through ``action_type=phonetic_reading``, same arguments.
 
 ## 2.8.1-fork (April 14 2023)